[FIX] mail, crm, crm_livechat: Assign to correct UTMs
The creation of lead is corrected when conversation happens via whatsapp
channel. After this commit, the correct UTM source and medium is set
after the lead is created.

Additionally, /lead command was moved to crm module, which makes more
sense as leads can come from different sources and livechat installation
shouldn't be required to use this functionality.

task-3964124
