+++ addons/crm/__manifest__.py
@@ -84,13 +84,11 @@
         ],
         'web.assets_unit_tests': [
             'crm/static/tests/mock_server/**/*',
-            'crm/static/tests/crm_test_helpers.js'
+            'crm/static/tests/crm_test_helpers.js',
+            'crm/static/tests/composer.test.js',
         ],
         'web.qunit_suite_tests': [
-            'crm/static/tests/**/*',
-            ('remove', 'crm/static/tests/tours/**/*'),
-            ('remove', 'crm/static/tests/mock_server/**/*'),
-            ('remove', 'crm/static/tests/crm_test_helpers.js')
+            'crm/static/tests/legacy/**/*'
         ],
     },
     'license': 'LGPL-3',
