+++ addons/crm/static/tests/composer.test.js
