+++ addons/crm/static/tests/legacy/crm_kanban_progress_bar_mrr_sum_field_tests.js
