+++ addons/crm/static/tests/legacy/forecast_kanban_tests.js
