+++ addons/crm/static/tests/legacy/forecast_view_tests.js
