+++ addons/crm/static/tests/legacy/model_definitions_setup.js
